fix: rename shared-posts to shared-posts-page
