+++ src/app/pages/index.ts
@@ -5,4 +5,4 @@ export * from './profile-page';
 export * from './setting-page';
 export * from './user-input-page';
 export * from './writing-post-page';
-export * from './shared-posts';
+export * from './shared-posts-page';
