+++ src/app/pages/shared-posts-page.tsx
